refactor(spatial-ai): extract pipeline result models
